% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_predict.R
\name{ai_predict}
\alias{ai_predict}
\title{Function for predicting data with a trained AI}
\usage{
ai_predict(matched_dtm, trained_learner, verbose = FALSE)
}
\arguments{
\item{matched_dtm}{Matched document-term-matrix of the documents for which
the AI should #' predict data. Ideally this matrix is created via
\code{\link{match_into_model}}. The text_model must be the same as stored
in the \code{trained_learner} object.}

\item{trained_learner}{Trained learner for predicting the data. The object
must be generated via the function \code{\link{ai_train}}.}

\item{verbose}{\code{TRUE} if information on the progress should be printed
to the console.}
}
\value{
Returns a \code{matrix} with the predications for every document.
}
\description{
This function predicts new data based on a trained AI. The AI should be
created by the function \code{\link{ai_train}}.
}
