% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_train.R
\name{ai_train}
\alias{ai_train}
\title{Training of Artificial Intelligence}
\usage{
ai_train(
  basic_text_rep,
  text_model,
  learner,
  normalize = FALSE,
  n_performance_estimation = 30,
  ratio_performance_estimation = 0.75,
  category_name,
  filter = "jmim",
  filter_ratio = 1,
  verbose = FALSE
)
}
\arguments{
\item{basic_text_rep}{Textual representation from which the AI should learn.
The object should be generated with the function
\code{\link{create_basic_representation}}.}

\item{text_model}{The model of the texts that should be used for training AI.
This object is must be generated with \code{\link{create_text_model}}.}

\item{learner}{Learner used for modeling the AI. It must be created with
mlr3.}

\item{normalize}{\code{TRUE} if the independent and depended variables should be normalized
before training AI. See details for more information.}

\item{n_performance_estimation}{Number of iterations for estimating the
performance of AI via bootstrap.}

\item{ratio_performance_estimation}{Ratio for splitting the data into
training data and test data.}

\item{category_name}{Name of the target variable that should be predicted.}

\item{filter}{Name of the filter that should be used.}

\item{filter_ratio}{Amount of features that should be included
in training the AI after applying a filter. For including all features set 1.0.}

\item{verbose}{\code{TRUE} if information on the progress should be printed
to the console. \code{FALSE} if not desired.}
}
\value{
Function returns an object of class \code{List} that contains
the trained learner and all necessary components for predicting new data.
}
\description{
Function for training an AI with several configurations and options
provided by several packages (e.g. mlr3, text2vec).
}
\details{
\itemize{
\item{\code{method:}}{Currently the function supports four methods.
\strong{"DTM"} forces that AI to learn based in the document-term-matrix.
\strong{"LSA} forces that an LSA is performed based on the initial text
representation. AI uses the solution of the LSA.
\strong{"LDA"} forces that an LDA is performed based on the initial text
representation. AI uses the solution of the LDA.
\strong{"GLOVE"} performs the estimation of global vectors based on the initial
text representation. The result is used for learning by weighting the the
global vectors with the word counts of the document-term-matrix.}
\item{\code{normalize: }}{If \code{TRUE} all data is normalized by applying the
following formula to all columns of the data.frame:
\deqn{(0.95*(x - min(x)) / (max(x) - min(x)))+0.01}}
}
}
