% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Text_Representation.R
\name{create_basic_representation}
\alias{create_basic_representation}
\title{Creates basic text representations}
\usage{
create_basic_representation(path_language_model, data, verbose = TRUE)
}
\arguments{
\item{path_language_model}{Path to the language model which should be used
for analyzing the data with help of the package udpipe.}

\item{data}{Object containing the concrete texts.Ideally this object is
created via the function \code{readtext} from the readtext package.}

\item{verbose}{\code{TRUE} if information on the progress should be printed
to the console. \code{FALSE} if not.}
}
\value{
Returns a \code{List} with the following two components:
\itemize{
\item{\code{dtm: }}{Document-term-matrix of class \code{dfm} created with
the package quanteda.}
\item{\code{fcm: }}{Feature-co-occurence-matrix of class \code{fcm} created
with the package quanteda.}
}
}
\description{
Function for generating the document-term-matrix and the feature-co-
occurrence-matrix for training and predicting.
}
