% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R
\name{create_filter_plot}
\alias{create_filter_plot}
\title{Create plot of filter values}
\usage{
create_filter_plot(trained_learner)
}
\arguments{
\item{trained_learner}{Trained learner for predicting the data. The object
must be generated via the function \code{\link{ai_train}}.}
}
\value{
Returns a plot of the relevant values.
}
\description{
Function for creating a visualization of the relevant filter values.
}
