% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_representations.R
\name{create_text_model}
\alias{create_text_model}
\title{Creates a model of texts}
\usage{
create_text_model(
  basic_rep,
  method = "DTM",
  min_termfreq = 2,
  min_docfreq = 2,
  dim = 10,
  max_iter = 5000,
  verbose = TRUE
)
}
\arguments{
\item{basic_rep}{Basic textual representation generated with the function
\code{\link{create_basic_representation}}.}

\item{method}{Method for building an advanced text representation as input
for the AI. See details for more information.}

\item{min_termfreq}{\code{Int} for determining the minimal appearance of a
term in the whole data set.}

\item{min_docfreq}{\code{Int} for determining the minimal number of documents
that must contain a term. If the term is used in a lower number of documents
it is excluded from the analysis.}

\item{dim}{\code{Int} representing the number of modeled dimensions.
It is only used in combination with methods "LDA","LSA", and ' "GLOVE".}

\item{max_iter}{Maximum number of iterations allowed for "LDA", "LSA", and
"GLOVE".}

\item{verbose}{\code{TRUE} if information on the progress should be printed
to the console. \code{FALSE} if not desired.}
}
\value{
Returns a \code{list} which contains the method used, the text model
\code{transformation_matrix} for further analysis, and a matrix for matching
new data into the structure of the
generated text model(\code{match_f_v_trim}).
}
\description{
Function for creating a textual model that provides the structure for
preparing data for training AI and prediction new data.
}
\details{
Currently the function supports four methods.
\itemize{
\item{"DTM" }{forces that AI to learn based in the document-term-matrix.}
\item{"LSA" } {frces that an LSA is performed based on the initial text
representation. AI uses the solution of the LSA.}
\item{"LDA" } {forces that an LDA is performed based on the initial text
representation. AI uses the solution of the LDA}
\item{"GLOVE" } {performs the estimation of global vectors based on the initial
text representation. The result is used for learning by weighting the the
global vectors with the word counts of the document-term-matrix.}
}
}
