% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_representations.R
\name{match_into_model}
\alias{match_into_model}
\title{Matching new data into the text model}
\usage{
match_into_model(dtm, text_model, verbose = FALSE)
}
\arguments{
\item{dtm}{Document-term-matrix generated containing the new data. Object
is ideally generated with the function \code{\link{create_basic_representation}}.}

\item{text_model}{Textmodel in which the new data should be fitted. Object is
ideally created with the function \code{link{create_text_model}}.}

\item{verbose}{\code{TRUE} if information on the progess should be printed
to the console. \code{FALSE} if not desired.}
}
\value{
Returns a \code{matrix} which represents the new data at the fundation
of the text model.
}
\description{
Function for matching new textual data into an existing text model.
}
